moved api key to server
For reference, this wasn't due to a security concern. The API key previously on the page was strictly a read-only key that was only able to read events from a specific account, and those events are already all on piginapickle.com publicly anyway. In the worst case (if someone were to feel the need to abuse this key), we'd reach our API monthly quota and the calendar 'Special Events' wouldn't load on the site. Now, with the key on the server, we can handle our rate limiting however we'd like to. Wasn't really a concern before, but now this just buys us some more control.